Style correction:
Previously it used dark everywhere.
Now the color that is used everywhere is primary, this makes more sense

Although, in order to keep the same visual appearance,
the primary has been assigned a custom color, equal to that of the dark.

This leaves dark and primary as the same,
and the default color of primary is technically gone.